assignment_materials/04.agentic_workflow/pyproject.toml
[project]
name = "04-agentic-basic-threatintelligent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.5.0",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "markdown>=3.10.2",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.2.1",
    "scalar-fastapi>=1.8.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
    "weasyprint>=68.1",
]
